--- a/data.sql
+++ b/data.sql
@@ -78,3 +78,10 @@ VALUES (1, 1, '2020-05-24'),
         (9, 2, '2020-08-03'),
         (10, 3, '2020-05-24'),
         (10, 1, '2021-01-11');
+
+-- -- Added 17 times each one to pass 1000ms mark
+-- This will add 3.594.280 visits considering you have 10 animals, 4 vets, and it will use around ~87.000 timestamps (~4min approx.)
+INSERT INTO visits (animal_id, vet_id, date_of_visit) SELECT * FROM (SELECT id FROM animals) animal_ids, (SELECT id FROM vets) vets_ids, generate_series('1980-01-01'::timestamp, '2021-01-01', '4 hours') visit_timestamp;
+
+-- This will add 2.500.000 owners with full_name = 'Owner <X>' and email = 'owner_<X>@email.com' (~2min approx.)
+INSERT INTO owners (full_name, email) select 'Owner ' || generate_series(1,2500000), 'owner_' || generate_series(1,2500000) || '@mail.com';