# FEATURES

- Create email column to owners table
- Add values to email column and to visits table
- Performance test run using 3 queries promps. 

## SCREENSHOTS

### Result of the 3 promps before improving the performance
![Screenshot 2023-07-17 143947](https://github.com/enmanuelbayen/Vet-clinic-DB/assets/121642946/5214bf95-36a0-4bee-b0e9-2ad305a54034)

### 1st promp performance improve from 1160ms approx. -> 336ms approx.
>![Screenshot 2023-07-17 154846](https://github.com/enmanuelbayen/Vet-clinic-DB/assets/121642946/856cc233-6a23-426d-a619-65c137e77c0d)

### 2nd promp performance improve from 3k-5kms approx. -> 1.5kms approx.
>![Screenshot 2023-07-17 161238](https://github.com/enmanuelbayen/Vet-clinic-DB/assets/121642946/408cffdd-7c04-4eda-a912-940db5152f1e)

### 3rd promp performance improve from 1300ms approx. -> 1ms approx.
>![Screenshot 2023-07-17 161720](https://github.com/enmanuelbayen/Vet-clinic-DB/assets/121642946/de7af805-0c34-4c98-be40-425f1271b6eb)